backend/requirements.txt
# Web Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic[email]>=2.7.4
pydantic-settings>=2.1.0

# AI & RAG Components (Python 3.12 compatible)
llama-index-core>=0.10.30
llama-index>=0.10.30
llama-index-llms-langchain>=0.1.3
llama-index-embeddings-langchain>=0.1.2
langchain>=0.1.20
langchain-google-genai>=1.0.10
langchain-community>=0.0.38
google-generativeai>=0.8.3

# Content Extraction
youtube-transcript-api>=0.6.1
beautifulsoup4>=4.12.2
requests>=2.31.0
PyPDF2>=3.0.1
python-docx>=1.1.0
python-pptx>=0.6.23
pandas>=2.1.3
openpyxl>=3.1.2
pillow>=10.1.0
pytesseract>=0.3.10

# Language Detection & Translation
langdetect>=1.0.9

# Vector Database
chromadb>=0.4.18

# Email
fastapi-mail>=1.4.1

# Utilities
python-dateutil>=2.8.2
validators>=0.22.0

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2
